Updates eg.Config and eg.PersistentData
Changes the roll of eg.Config. instead of it being the main handler for the ConfigData it acts more like a subclass of PersistentData.

Created eg.PersistentData.PersistentDataBase which is what an eg.PersistentData subclass becomes. instead of eg.Config.Section

Added extra functionality to the base class. if you iterate through the class it yields key, value pares that it has stored skipping any keys that start with underscore. helps to minimize the code footprint. and also remove the use of iteritems as this is no longer available in python 3

Has each instance of the base handle it's own writing of information. and added checking to see if the class is to be deleted or if the class was loaded only to satisfy the requirements of debugging and can be removed on close.

moved any class checking into the base class under __setattr__ in the event a class gets added dynamically this will change that class to an instance of the base class. as this is a container for settings. it should not have any class instances stored in it dynamically.

Fixes issue when a plugin exists and is installed If the developer added a new attribute to the class and the plugin is upgraded it would not set the new attribute into place and when the attribute would get accessed from the plugin a traceback would occur. Unfortunatly due to having to iterate through each class and add any attributes that do not exist. this adds about 100ms to EG's startup time.

Added formatting of data where the line length is to long. or is a list or tuple that is to long for one line.

It now writes all configuration data to the config.py. not just the values that have changed. This was done because of possible changes to default values in the future. if EG is upgraded all the proper settings can be carried over to the new version regargless of any changes made to the default values.

The big goal to this was to change how the PersistentData and Config work in order to move them to the top of the Core.py file. by doing this the debugging level can get set one time depending on a command line switch or a svaed value in the config.py.