Fixes multiple things
Fixes the event names not showing in the dialog
Moved the loading of the file into eg.Document.Load this way if someone changes the tree file and they have different plugins this will load the plugins. still have to work out keeping the saved events between tree changes.

Fixes the finding the last used event in the add events dialog